RL_Agent/requirements.txt
absl-py==0.8.1
asgiref==3.2.7
astetik==1.9.9
astor==0.8.0
atari-py==0.2.6
backcall==0.1.0
cachetools==3.1.1
certifi==2019.9.11
chances==0.1.9
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.1
Django==3.0.5
dnspython==1.16.0
eventlet==0.25.1
future==0.18.2
gast==0.2.2
geonamescache==1.1.0
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
greenlet==0.4.15
grpcio==1.25.0
gym==0.15.4
h5py==2.10.0
idna==2.8
image==1.5.31
ipython==7.11.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
joblib==0.14.1
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kerasplotlib==0.1.6
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
monotonic==1.5
numpy==1.16.4
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandas==1.0.4
parso==0.5.2
patsy==0.5.1
pickleshare==0.7.5
Pillow==7.1.1
prompt-toolkit==3.0.2
protobuf==3.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyglet==1.3.2
Pygments==2.5.2
pyparsing==2.4.5
python-dateutil==2.8.1
python-engineio==3.10.0
python-socketio==4.4.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.3.3
seaborn==0.10.0
six==1.13.0
sklearn==0.0
sqlparse==0.3.1
stable-baselines==2.10.0
statsmodels==0.11.0
talos==0.6.6
tb-nightly==1.14.0a20190301
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019030115
tqdm==4.42.0
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.8
Werkzeug==0.16.0
wrangle==0.6.7
wrapt==1.11.2
